requirements.txt
pydoll-python
requests
random-user-agent
beautifulsoup4